Avoid copies of base::RepeatingCallback in media::ChunkDemuxer
media::ChunkDemuxer::SetParseWarningCallback simply forwards its
argument along to SourceBufferState::SetParseWarningCallback, though it
incurs a copy when doing so. A simple fix to this is to use std::move,
which this CL does.

A more complex case is media::ChunkDemuxer::SetTracksWatcher, which
takes its argument by const& and only incurs a copy in
SourceBufferState::SetTracksWatcher. Since base::RepeatingCallback is
essentially just a scoped_refptr, it should follow the style guide
recommendations for that type and pass by value when the callee is
guaranteed to make a copy. Therefore this CL updates those functions as
well.

Additionally, this CL extends the comments on
it clear that calling these functions is required before sending data to
the ChunkDemuxer.

media: :ChunkDemuxer::(SetParseWarningCallback|SetTracksWatcher) to make
Change-Id: Id3cca2599c144d3d49c5d3c66f403ebff4196022
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3088154
Commit-Queue: Will Cassella <cassew@google.com>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910640}
NOKEYCHECK=True
GitOrigin-RevId: cb415b85d25bc24accc131964e35ae7e4b1d9829