PaintCanvasVideoRenderer: Inline CopyVideoFrameSingleTextureToGLTexture
The functions CopyVideoFrameSingleTextureToGLTexture and
ImportVideoFrameSingleMailbox only have one caller. Inline them into
their caller.

CopyVideoFrameSingleTextureToGLTexture contains the admonistration that
"the caller must call SynchronizeVideoFrameRead after this operation",
but that lone caller does not do so. Fix that. Parameterize the type of
the interface to SynchronizeVideoFrameRead  so that it can be called on
both gpu::gles2::GLES2Interface and gpu::raster::RasterInterface. Add
some comments to draw attention to the importance of the function.

Bug: 1237100
Change-Id: I371f1e69d1c65e42e451435fb9de9b432f6bceeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3089013
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910959}
NOKEYCHECK=True
GitOrigin-RevId: f8f486b3653ca2c6535f46f62e96ac79a7d73516