PaintCanvasVideoRenderer: Clean up UpdateLastImage
Update the behavior of PaintCanvasVideoRenderer::UpdateLastImage to
early-out on a cache hit to improve readability (rather than have
the entire function be indented).

Be precise about calls to SynchronizeVideoFrameSingleMailbox. It is
to be called after a copy that directly references a VideoFrame.
If we're copying the VideoFrame into a cache, then do it after that
copy. If we're copying a cached texture into another texture, do it
only if that cached texture is referencing a VideoFrame.

The helper function SynchronizeVideoFrameSingleMailbox has a single
caller, inside UpdateLastImage. Merge it in.

Bug: 1237100
Change-Id: I61f07594bcebbc9693e17051b6157e958ec1bd14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3087353
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#910977}
NOKEYCHECK=True
GitOrigin-RevId: e042926372cd1c499aac259c63c0095e0b0d1a8f