media: V4L2VideoDecoder: fix the order between the Decode() and Reset()'s callback
Currently, the callback of Decode() is always trampolined once to the
working task runner. This makes it possible for the Reset() callback
to run before all the decode callbacks prior to the Reset() are
called, and this goes against the media::VideoDecoder interface.

This CL fixes the issue by trampolining the callback of Reset(). Now
the callbacks of both Decode() and Reset() always trampoline once, so
the calling order can be preserved.

BUG=b:192087509
TEST=arc.VideoDecodeAccelVDVDA.h264

Change-Id: Ic649cd8aa25937a654b3c3739d6d0bd128c0b11f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3084026
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#911595}
NOKEYCHECK=True
GitOrigin-RevId: 1c03b553d1bf085c6f0d1c2d484bf281dc4b3bdf