webcodecs: Fallback to software if hardware encoder fails
When the user doesn't insist on hardware encoding
(VideoEncoderConfig.hardwareAcceleration="no-preference" or not set)
we have flexibility of choosing encoder's implementation.

Currently if an accelerated encoder can be created by GPU factories,
we stick with this implementation regardless of how successful later calls
to Initialize() and Encode(). So if the accelerated encoder fails,
for example if all available platform encoders are already in use, the user
will get a broken encoder even if they don't care or know about acceleration.

This change introduces a VideoEncoderFallback, a proxy encoder that
delegates all work to an accelerate encoder and creates a software one
when the accelerated encoder fails during initialization
or later while encoding.

Bug: 1148593
Change-Id: Ibd61ef1b2073380c7a2a1069f39b84b6c05c2a3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3089237
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#912473}
NOKEYCHECK=True
GitOrigin-RevId: ca846b7f1543b928fc027300df54aef8ff47a08f