Remove DCHECK for is_capturing when stopping video capture
A DCHECK for the variable is_capturing_ gets triggered when StopStream()
is called. The default value of is_capturing_ is false and the only time
it is set to true is at the end of StartStream(), but any time
SetErrorState() gets called the boolean variable is_capturing_ is set to
false and since StopStream() will get called after an error has been
detected and SetErrorState() was called, leading to this DCHECK
triggering and the browser crashes.

Suggested fix is to remove the DCHECK in StopStream() since it is
supposed to deallocate resources if an error occur i.e. when the web
camera is already in use.

Change-Id: I0e63738bce2879031b048278de7533a7a558abb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3043975
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: mark a. foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#913085}
NOKEYCHECK=True
GitOrigin-RevId: fcd0e6c8e860cc17cd930523a9fe1a7f9f7cbd6b