Implement RasterInterface::ConvertRGBAToYUVAMailboxes
In media::CopyRGBATextureToVideoFrame, add the client-side code to use
the RasterInterface method when there is no client-side GrContext.

Implement RasterDecoderImpl::DoConvertRGBAToYUVAMailboxesINTERNAL. Most
of the function body has been made to be shared with the YUV->RGB
conversion function. It may be that more code can be shared there.

In SharedImageRepresentationSkiaImpl::BeginWriteAccess, use
kAlpha_8_SkColorType instead of kGray_8_SkColorType, because the latter
is not renderable.

Bug: 1206168
Change-Id: I54b1714200816036d94d1d2b5d34dd1acffdb947
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3097373
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#914885}
NOKEYCHECK=True
GitOrigin-RevId: d18b932d56ab069f16c2c219a9e30b57953497b5