media: Add CreateVaapiImageProcessorWithInputCandidates().
This CL adds CreateVaapiImageProcessorWithInputCandidates() to allow the
ImageProcessorFactory to create a VA-API-based ImageProcessor given a
list of input format candidates.

This is in preparation to let the VideoDecoderPipeline do scaling using
an ImageProcessor. Thus, this CL refines the parameters of
ImageProcessorFactory::CreateWithInputCandidates() to allow passing an
input visible rectangle and the desired output size.

Test: None.
Bug: b:190545350
Change-Id: I7cd761cb90bb2ed0b77b614b2a6a0c4a7e093eec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3083172
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: J Kardatzke <jkardatzke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#916159}
NOKEYCHECK=True
GitOrigin-RevId: 7d50c01ba9b9650a9f6116dca43f713378a36627