media/AllocateGpuMemoryBufferHandle(): Remove bad DCHECK.
This CL removes a DCHECK from AllocateGpuMemoryBufferHandle() which was
introduced to document that we only use the
nullptr-GpuMemoryBufferFactory path for encoding.

With the recent CL:3089495, we can also use this path for the decoder:
specifically, when we need to scale decoded frames using an
ImageProcessor. The pool of buffers for the decoder will be allocated
with this nullptr-GpuMemoryBufferFactory path. The main implication is
that it won't try to create DRM/KMS framebuffers for the allocated
buffers which is fine because those buffers won't get promoted to
overlays. They're just used as the input for the ImageProcessor.

Test: None
Bug: b:190545350
Change-Id: Idbc01bedb7ff8d5447826ed02d679378259d305e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3149691
Auto-Submit: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: J Kardatzke <jkardatzke@chromium.org>
Reviewed-by: J Kardatzke <jkardatzke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919373}
NOKEYCHECK=True
GitOrigin-RevId: 92a17d0727ef2d06baa7ac3404f4300e80bf29fe