media: Not use AAC extradata in FFmpegAudioDecoder
Previously AAC extradata is only enabled on Android. Recently we landed
a change to enable it on all platforms [1].

However, there is an issue with hls.js which could be generating AAC
contents with wrong AAC extra data (ASC), causing playback failure.
Since it's unclear how prevalent this is in the wild. For backward
compatibility, we do not pass the AAC extradata to FFmpeg, which is the
same behavior as before.

In the future, after we understand more about the hls.js issue and its
impact, we may be able to remove this hack.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/3044929

Bug: 1245123
Test: The test content in the bug plays.
Change-Id: I2a96eaa5a466d6e4fc99d6625a0dbe230abe29b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3149840
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919896}
NOKEYCHECK=True
GitOrigin-RevId: 589eac300fdfc91c4d0a2af7f63482d3760868fe