media: Report Clear Lead usage in MediaLog
In addition to report Clear Lead to UKM, also report it to MediaLog so
it'll show up in media-internals for easier local debugging. We only
report when there is clear lead because {Audio|Video}DecoderConfig is
already reported to MediaLog which contains the clear/encrypted info.

Bug: 1225501
Change-Id: Ibacabe21e3b1f23bbba3b95e0df9936d6336730b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3151393
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920425}
NOKEYCHECK=True
GitOrigin-RevId: e873467ce36a7e5c7c975d46cfb80eece8525a54