vaapi: fix use-after-frees
We're passing pointers as a part of a GPU buffer. Those need to be
processed and properly serialized before we can free those pointers. The
processing/serialization takes place after we unmap the buffer. Hence,
we need these objects to remain alive on the stack until after we send
the buffer.

Bug: b:199449438
Change-Id: Id76091e1dc747df8e3558aa737053d7678cce30b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3165557
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922251}
NOKEYCHECK=True
GitOrigin-RevId: eb2d87129010fba1ac60951a3af8f88ab14cbc8c