Fix possible UAF during AAudio errors
Recent changes to the AAudioOutputStream added a destruction helper,
which guarantees that the AAudio stream would not be used after it's
freed, in the data callback.

This CL updates the error callback to also use the destruction helper,
to avoid potential UAF, if the AAudioOuputStream is already destroyed
when the error callback runs.

Bug: 1207941
Change-Id: I019e3594ce1672cf0f11c80d3d2940c23776e101
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3166999
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922749}
NOKEYCHECK=True
GitOrigin-RevId: 80a7b7079aaf46bb027eeb58056184831f0732e4