V4L2(S)VDA: Set right visible rect for image processor input.
This CL fixes both the V4L2SliceVideoDecodeAccelerator and
V4L2VideoDecodeAccelerator to ensure the input for the image processor
has the expected visible rectangle.

The reason this is needed is that CL:3171767 introduced a check to
ensure that the visible rect of the frame input to the image processor
is the same as the visible rect used to configure the image processor.

Before the fix in this CL, the visible rectangle of the video frame
would just be the coded size. That's because the video frame is created
with V4L2Buffer::CreateVideoFrame() which uses format_.fmt.pix_mp.width
and format_.fmt.pix_mp.height [1]. With the fix in this CL, we wrap that
video frame into another that has the correct visible size. Note that
the correct visible size is tracked in the
V4L2(Slice)VideoDecodeAccelerator. This is more or less consistent with
how the V4L2VideoDecoder does it [2].

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/v4l2/v4l2_device.cc;l=318-321;drc=4c42559f995f5a88da6dc55642eaf5244693e033
[2] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/v4l2/v4l2_video_decoder.cc;l=755-756;drc=cab37e297f59644b0264f926093639f69580b7ad

Bug: b:195351653, b:190545350, b:201160897, b:201341614
Test: tast run $IP video.DecodeAccel.* on hana/R96-14250.0.0
Test: tast run $IP video.DecodeAccel.* on kukui/R96-14250.0.0
Change-Id: I98ebe4fdc9603e3bff52505215761228c1547134
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3198826
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#929027}
NOKEYCHECK=True
GitOrigin-RevId: 693314f8d2ed64c270d6b27b287b293ef91de349