media/gpu/vaapiVED: Introduce EncodeResult
VaapiVideoEncodeAccelerator carries the resource that requires
encoding with EncodeJob until the end. While awaiting a buffer
that encoded is written into after encoding is complete, most
variable and functions are not needed anymore.
This introduces EncodeResult so that we can keep the encoded
data only after encoding. As the first step, it just composes
EncodeJob in this CL. EncodeJob will be removed in the future.

Bug: b:202337642
Test: media_unittest
Test: video_encode_accelerator_tests
Change-Id: Iace58f6785686978071794737a04aedae3e91250
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3213973
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#930043}
NOKEYCHECK=True
GitOrigin-RevId: 2116f9e933b34f3ff970923781eb5fa390cea214