media: pass CroStatus to V4L2VideoDecoderBackend::OnChangeResolutionDone()
Originally, we pass a boolean value to OnChangeResolutionDone() to
indicate if the resolution change process is done successfully or not.
However, the process might be aborted by the reset request. In this
situation, OnChangeResolutionDone() is not called forever.

This CL changes to pass a CroStatus to OnChangeResolutionDone().
CroStatus::Codes::kResetRequired indicates the aborted situation.
Also, we always call OnChangeResolutionDone() for each resolution
change process (except the instance is about to be destroyed). Then
the backend implementation could handle the aborted situation
individually.

BUG=b:192523692
TEST=tast.arc.VideoDecodeAccelVDVDA.h264
TEST=tast.video.DecodeAccelVD.h264

Change-Id: Ia480ade77d615a8d2dd221b049a49bf670b490fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3180501
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#933273}
NOKEYCHECK=True
GitOrigin-RevId: d79a96cd7a5e3ce9ab80a1ccd4307c4a7425c45a