[Fuchsia] Use dedicated thread for AudioCapturer.
Added a new thread that's used to receive packets from AudioConsumer and
pass them to CapturerCallback::Capture(). This ensures that incoming
packets are handled as soon as possible and are not blocked on the tasks
that run on the main renderer thread.

Bug: 1263276
Change-Id: I50ed8673c48f79e16d3e6b8ba6928967682daa6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3244002
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#935146}
NOKEYCHECK=True
GitOrigin-RevId: b136d473b7f25010c85fb2928123df31c363c0d6