VideoFrameValidator: destroy VideoFrameMapper on validator thread.
This CL makes the VideoFrameValidator destroy its |video_frame_mapper_|
on the thread in which it creates it, i.e., the
|frame_validator_thread_|.

This is needed specifically for the VaapiDmaBufVideoFrameMapper
implementation which owns a VaapiWrapper, and because of CL:2793005, the
VaapiWrapper must be created, used, and destroyed on the same sequence.

The |corrupt_frame_processor_| is also destroyed on the validator thread
before the |video_frame_mapper_| in case it has references to mapped
VideoFrames.

Bug: b:183515581
Test: ./video_decode_accelerator_tests
Change-Id: I2e28fa1fa027c493ba5aca6c6a2e1ef4d055ebad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3246247
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#935492}
NOKEYCHECK=True
GitOrigin-RevId: 99d9a032ec597382760f76d700c48483bd8d2a0d