Don't enforce VaapiWrapper seq. affinity in VaVDA.
The VaapiVideoDecodeAccelerator is a legacy class that is known to use
at least one of its VaapiWrappers from multiple threads. This violates
the restriction introduced by CL:2793005. Unfortunately, refactoring
this class to use the VaapiWrapper on a single sequence is non-trivial.

Thus, to avoid making major changes to a class that's on its way out, we
allow it to create VaapiWrapper instances that don't enforce sequence
affinity.

Bug: b:204394724, b:183515581
Test: tast run ${IP} video.Play.h264_hw_alt on volteer
Change-Id: I4fcfe0a9608a61652c109f6bb16ab169d4b8e513
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3252771
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#936566}
NOKEYCHECK=True
GitOrigin-RevId: 69171ea7843f8746233daaea95ea4e63aedf6d51