Add enable_hls_demuxer GN arg
This CL adds the `enable_hls_demuxer` arg to media_options.gni, which
defaults to false. There is currently no code landed behind this arg,
but having this arg checked in simplifies the process of protoyping. For
a preview of code which will be landed in future CLs behind this arg,
please see the prototype HLS demuxer:
https://chromium-review.googlesource.com/c/chromium/src/+/3096950

Bug: 1266991
Change-Id: Ic7c232750b7973f6e199c561da0053a72195b164
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3262414
Commit-Queue: Will Cassella <cassew@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938658}
NOKEYCHECK=True
GitOrigin-RevId: 0b46a2c725cfea5a5196793c9eb6926ccd2758b1