[CWR] Add Playback Controls to CastStreamingRenderer
This CL adds a new Mojo Pipe To Control playback of the
CastStreamingRenderer, as well as the infrastructure required to
use this new pipe - both in the //media/cast/receiver renderer-process
code where the receiver of this pipe will be located and in the
//chromecast/cast_core browser-process code where the remote will be.

Future changes will implement further playback controls as required
for remoting, while the current CL only adds those to start playback
as required for mirroring.

For the short term, this mojo pipe (or something similar with
comparable infrastructure - see: go/cwr-mirroring-test ) is required
for starting a new mirroring session. A deadline is approaching, so
I would prefer to iterate on the remoting use case in a follow up CL
rather than making that the focus of this CL. That being said, for
details on the long-term remoting use, see: go/remoting-in-the-cwr

Bug: b/187332037, b/190564826, b/182429113
Change-Id: I7ca5d33d4fe82cd1a1d1576ee1529e4d12831a24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3227722
Commit-Queue: Ryan Keane <rwkeane@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#938916}
NOKEYCHECK=True
GitOrigin-RevId: fdc06b00b7de8ba1d09485bfeb8c3294d19d0c15