media/gpu/test/EncodedDataHelper: Don't group ivf frames unless VP9 or AV1
EncodedDataHelper groups ivf frames whose timestamps are same.
The grouping is required for frames in spatial layers stream.
This lets DecoderBuffer contain multiple frames and inputs it
to a VideoDecoder. This causes a problem in vp8 parsing because
vp8 parsing process regards one buffer as one frame.

This CL changes EncodedDataHelper so that it doesn't group unless
the codec is VP9 and AV1.

Bug: b:205723249
Test: video.DecodeAccel.*
Test: video.ChromeStackDecoding.vp8_inter

Change-Id: I9def7e55a426da7c5cee6ede6654dc38e6abac5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3275132
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#941069}
NOKEYCHECK=True
GitOrigin-RevId: 482d54dd4be71aced82e27c413f908265e7f078f