media/gpu/v4l2: file descriptor for media device
This CL changes media device name from /dev/media0 to /dev/media-dec0.
This is because both /dev/media0 and /dev/media1 can be used for the
purpose even with the same device. This media device is linked to
/dev/media-dec0. Note that this is also platform dependent, so we might
need to add something else for another platform in the future.

Bug: b:203004039
TEST: v4l2_stateless_decoder --video crowd_run_256X144_fr15_bd8_8buf_l1_20210312.ivf on asurada
Change-Id: Ia5efd5792c00baf3bfa97e1c99429b95e60e5f3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3296386
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Steve Cho <stevecho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944605}
NOKEYCHECK=True
GitOrigin-RevId: 3b0a9a632da7f01cc9e2c9b6de12a0a60a8e77cd