HLS Manifest Parser: Add line-level lexical item extraction
This CL adds the bit of the HLS parser that is capable of extracting
line-level lexical items from a manifest (i.e. tags or URIs). To allow
this function to be simple and recoverable, any validation beyond EOL
checking is deferred to more contextual parts of the parser. This CL
adds unit tests and fuzzing to verify correctness and API guarantees.

Bug: 1266991
Change-Id: I40b1642a0618896fa63a0a243f62f0deb8d6777d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3312628
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947825}
NOKEYCHECK=True
GitOrigin-RevId: 5294a04692f365141f210d59f757df58f139ff4b