media: Make VideoSourceProvider usable on chromeos-linux
Using VideoSourceProvider on chromeos-linux builds
(generally knows as the emulator) isn't possible due to
the fact that it relies on VideoCaptureDeviceFactoryChromeOS
which uses the camera hal that is built only for chrome os
images installed on actual devices.

This CL creates VideoCaptureDeviceFactoryLinux instead
when we're not running on a device, so that using the
emulator is possible to test camera-related features.

This will be a huge productivity boost for many of our
engineers.

`VideoSourceProvider::GetSourceInfos()` is able to get the
camera devices correctly on the emulator.

Bug: 1276474
Test: Tested manually that calling
Change-Id: I47e05be94e5f9cff6d09d99c63d9f08d0a7dc70c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3311677
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Tommi <tommi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948058}
NOKEYCHECK=True
GitOrigin-RevId: 082364cc22325d71b075e62d521814e12982e62f