HLS: Add SourceString and SourceLineIterator utilities
This CL introduces two utilities to replace the previous pattern of
passing `size_t line_number` and `base::StringPiece` around by mutable
reference: SourceLineIterator (which represents a line-based iterator of
an entire manifest) and SourceString (which represents a substring of a
single line from a manifest). This pulls some of the responsibility out
of `GetNextLineItem`, as was suggested in the previous HLS CL.

These utilities have been placed in the `hls/parse_context.h` header, as
there are other parsing-context related structures which will be added
here in the future.

Bug: 1266991
Change-Id: Ic51dd3501eafe7d45f4fe9ce6af0020921578cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3319782
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949213}
NOKEYCHECK=True
GitOrigin-RevId: c29ffc1d39d1bc8fd84b9d03696bf6809faa90ad