VaPicNativePixmapOzone: don't use Ozone in ImportGpuMemoryBufferHandle()
We don't really need to import the incoming GpuMemoryBufferHandle into
Ozone in order to use it for video decoding. This CL wraps the
NativePixmapHandle from the incoming GpuMemoryBufferHandle into a
NativePixmapDmaBuf and passes that to
VaapiPictureNativePixmapOzone::Initialize().

We do add a CanImportGpuMemoryBuffer() utility function to attempt
importing the incoming handle into minigbm which should keep some of the
checks that the Ozone path had before this CL.

This removal of Ozone as a dependency of ImportGpuMemoryBufferHandle()
is needed for out-of-process video decoding: in a video decoder process,
we don't initialize Ozone, so the code before this CL would crash.

Bug: b:195769334
Test: arc.VideoDecodeAccel.h264_vm on volteer
Change-Id: Idba9305b56bb31878bc115baab7f60eaf5a8071c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3316246
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949940}
NOKEYCHECK=True
GitOrigin-RevId: c9a3bec6c561e2a60757af4a8f765525e29c7f43