media/gpu/v4l2/V4L2JEA: Request active markers for V4L2_PIX_FMT_JPEG
The upstream kernel API does not specify which markers or segments are
present for a V4L2_PIX_FMT_JPEG image, only that it is a valid JPEG
stream. If some markers are desired, the proper way to request them
is to use the V4L_CID_JPEG_ACTIVE_MARKER control.

Use the control to request APP0, DQT, DHT segments to be present. APP0
is requested for the image to be JFIF compliant. DQT and DHT segments
are required for the image to be decodable.

Since not all drivers implement this control, nor implement it in the
same way, we ignore any errors and simply bail out.

BUG=b:206738124
TEST=Run jpeg_encode_accelerator_unittests on kevin-kernelnext with
     dev_debug=2, check dmesg for V4L2_CID_JPEG_ACTIVE_MARKER

Change-Id: I582fa670d634b6b2d7ffe29d247d8a951eede804
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297373
Auto-Submit: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950093}
NOKEYCHECK=True
GitOrigin-RevId: 2f672b9d84bdf866140f79d0c418df5915a38d81