Add ARC HW protected video decoding support for AMD
This enables the video decoding path in Chrome for the HWDRM system in
ChromeOS on AMD platforms when used with ARC. In this mode, there is a
special marker at the beginning of the video data that is followed by
a size and protobuf. The protobuf contains additional details needed
to handle the decoding. This then feeds into the normal
VideoDecoderPipeline which will handle everything else.

There will be a follow up CL which handles setting the quads to be for
HW protected video. This feature is not enabled until we turn on
USE_ARC_PROTECTED_MEDIA in the Chrome build config for ChromeOS.

BUG=b:203821839
TEST=Works with full prototype, existing playback still works

Change-Id: I76006d37cb7454e9f6ac560858ac9c772ce35225
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3331180
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/main@{#952092}
NOKEYCHECK=True
GitOrigin-RevId: 37e25e5bde1f1cd73749e5d6b6fdfea5eeb5eaee